Added feature "getpointvalue", which supports querying a location only by x and y. X and Y are in the coordinates in the requested projection space.

+++ adagucserverEC/CServerParams.h
@@ -15,7 +15,7 @@ class CServerParams{
   public:
     double dfResX,dfResY;
     int dWCS_RES_OR_WH;
-    int dX,dY;
+    double dX,dY;
     CT::string *WMSLayers;
     CT::string Format;
     CT::string InfoFormat;
